revert the refactor, split LCA into two rules
